tests/fixtures/frameworks/django_app/requirements.txt
Django==5.0.0
djangorestframework==3.15.0
